+++ src/uel/errors/throwexception.py
@@ -2,7 +2,8 @@ from os import _exit
 from sys import stderr
 
 from uel.errors.uelbaseexception import UELBaseException
-
+from uel.errors.uelbuildtimeexception import UELBuildtimeException
+import objprint
 STATUS = 1
 
 __all__ = ["ThrowException"]
@@ -11,8 +12,4 @@ __all__ = ["ThrowException"]
 class ThrowException:
     @staticmethod
     def throw(e: UELBaseException) -> None:
-        stderr.write(str(e))
-        stderr.write('\n')
-        stderr.flush()
-
-        _exit(1)
+        raise e