+++ src/uel/errors/ueerror_setstring.py
@@ -1,0 +1,4 @@
+import sys
+
+def UEErr_SetString(exce):
+    print(exce, file=sys.stderr)