+++ src/uel/errors/uelbuildtimeexception.py
@@ -1,8 +1,28 @@
 from uel.builder.position import Position
 from uel.errors.uelexception import UELException
+import linecache
+from uel.colors import RED, RESET
+
+import wcwidth
 
 __all__ = ["UELBuildtimeException"]
 
+def hint(line, col, offset):
+    hinted = " " * offset
+    
+    for char in line[0:col]:
+        width = wcwidth.wcwidth(char)
+        if width < 0:
+            width = 0
+        elif width == 0:
+            width = 1
+        hinted += " " * width
+    hinted += "^"
+    
+    if not line:
+        return "\n"
+    
+    return "\n" + (" " * offset) + line + hinted + "\n"
 
 class UELBuildtimeException(UELException):
     def __init__(self, error_message: str, pos: Position):
@@ -13,5 +33,7 @@ class UELBuildtimeException(UELException):
 
     def __str__(self) -> str:
         oes: str = super().__str__()
-        pos_string = f"{self.file}, {self.line}:{self.column}\n"
-        return pos_string + oes
+        line = linecache.getline(self.file, self.line)
+        pos_string = f"File {repr(self.file)}, line {self.line},\n"
+        return f"""\
+{pos_string}{hint(line, self.column, 2)}{oes}"""